Separate email template for mobile code login, login with code only on mobile
