+++ back/users/mail-mjml/require_login_template_mobile-replace.json
@@ -1,0 +1,3 @@
+{
+  "https://dailyidea-serverless-email-static-prod.s3.amazonaws.com": "{{ BUCKET_URL_PREFIX }}"
+}
