+++ back/users/mail-mjml/require_login_template_mobile.mjml
@@ -1,0 +1,32 @@
+<mjml>
+  <mj-head>
+    <mj-font name="Dancing Script" href="https://fonts.googleapis.com/css2?family=Dancing+Script&display=swap" />
+  </mj-head>
+  <mj-body background-color="#FCFCFC">
+    <mj-section background-color="white" padding-top="0" padding-bottom="0">
+      <mj-column>
+        <mj-image width="170px" src="https://dailyidea-serverless-email-static-prod.s3.amazonaws.com/email/images/Group18.png"></mj-image>
+      </mj-column>
+    </mj-section>
+    <mj-section background-color="white" padding-bottom="0">
+      <mj-column>
+        <mj-text padding-top="0" font-family="helvetica" font-size="18px" color="#2d2d2d">
+          <p>Your login code is:</p>
+          <p style="font-size: 20px; text-align: center;">{{code}}</p>
+        </mj-text>
+      </mj-column>
+    </mj-section>
+    <mj-section background-color="white" padding-top="0" padding-bottom="0">
+      <mj-column>
+        <mj-text padding-top="10px" padding-bottom="0" padding-right="0" font-family="helvetica" color="#2d2d2d" font-size="18px" line-height="22px">
+          <p>Return to your login page and enter your code to continue<p>
+          <p>This code will expire in 15 minutes and can be used only once.</p>
+          <p>
+            Lost your login page? Paste this link into your browser to continue. 
+            <a stlye="color: #2d2d2d;" href="https://{{ DOMAIN_NAME }}/auth/login/confirm?email={{email}}">https://{{ DOMAIN_NAME }}/auth/login/confirm?email={{email}}</a>
+          </p>
+        </mj-text>
+      </mj-column>
+    </mj-section>
+  </mj-body>
+</mjml>
